// server/types/offer.types.ts
type Offers = {
	id?: string
	date?: string
	photos: string
	title: string
	sale_status: string
	price: string
	developer_id: string
	type: string
	areas: string
	bed: string
	bath: string
	delivery_date: string
	finished: string
	location: string
	down_payment: string
	installment: string
	description: string
	active: string | boolean
	category: string
}

export default Offers